+++ app/models/computed_rule/exact_count_of_style.rb
@@ -6,9 +6,10 @@ module ComputedRule
 
     def self.build(house:, sections: Section, feature:)
       section = sections.random
+      style = Style.new(feature)
       count = house.count_styles(style: feature, section: section)
       rule = create
-      rule.text = "The #{section.name} must contain exactly #{count} #{feature} #{"object".pluralize(count)}"
+      rule.text = "The #{section.name} must contain exactly #{count} #{style.display} #{"object".pluralize(count)}"
       rule.save
       rule
     end