+++ app/models/empty_lamp.rb
@@ -1,3 +1,7 @@
 class EmptyLamp < Lamp
   include EmptyMixin
+
+  def icon
+    "<span class=\"icon-lamp\"><\/span>".html_safe
+  end
 end