+++ app/models/empty_wall_hanging.rb
@@ -1,3 +1,7 @@
 class EmptyWallHanging < WallHanging
   include EmptyMixin
+
+  def icon
+    "<span class=\"icon-wall-hanging\"><\/span>".html_safe
+  end
 end