+++ app/models/wall_hanging.rb
@@ -24,11 +24,15 @@ class WallHanging < Furnishing
   end
 
   def long_description
-    "A #{color_obj.display} #{style} wall hanging#{icon}".html_safe
+    "A #{color_obj.display} #{style_obj.display} wall hanging#{icon}".html_safe
   end
 
   def color_obj
     @color ||= Color.new(color)
+  end
+
+  def style_obj
+    @style ||= Style.new(style)
   end
 
   def icon